chore: move server.GenerateCoinKey and server.GenerateSaveCoinKey to testutil
